feat(launch): Rename Local Runners Continued (#3800)
* identfy change locations

* passing in launch queue

* breakpoints receiving commands from agentheartbeat

* sweep obj id

* Use launch queue

* modify scope of run queue

* modify run queue scope

* Sweep adds agent to launch runqueue

* uri is local path

* move further down to get project

* pass project explicitly

* some renames, catch bad uri

* remove breakpoints

* format

* check convert legacy sweep

* launch queue in upsert sweep

* launch sweep controller via cli

* boilerplate daimyo

* loading and seperate dependencies

* format

* more progress

* more progress

* TODOs, pending and start state

* legacy sweep run

* progress

* remove sweep from abc daimyo

* format

* cleanup push to run queue

* progress

* sweep id on upsert

* Sweep command

* no double upsert

* default entity

* pprint that mofo

* why must it be so

* its its aliive

* run over command

* cleanup

* todos

* get run state

* dataclass and enums

* threaded job polling

* remove threads, run id

* consolidation

* cleanup

* demo

* rename to scheduler

* format

* cleanup

* more cleanup

* cleanup 3

* cleanup, tests

* tests

* remove

* revert

* cli changes

* container

* utility test

* mypy

* LaunchError on invalid URI
