+++ tests/tests_launch/test_launch.py
@@ -1014,7 +1014,7 @@ def test_launch_local_docker_image(live_mock_server, test_settings, monkeypatch)
         "wandb.sdk.launch.builder.build.docker_image_exists", lambda x: True
     )
     monkeypatch.setattr(
-        "wandb.sdk.launch.runner.local._run_entry_point",
+        "wandb.sdk.launch.runner.local_container._run_entry_point",
         lambda cmd, project_dir: (cmd, project_dir),
     )
     api = wandb.sdk.internal.internal_api.Api(
@@ -1157,7 +1157,7 @@ def test_launch_local_cuda_command(
     live_mock_server, test_settings, monkeypatch, mocked_fetchable_git_repo
 ):
     monkeypatch.setattr(
-        "wandb.sdk.launch.runner.local._run_entry_point",
+        "wandb.sdk.launch.runner.local_container._run_entry_point",
         lambda cmd, _: cmd,
     )
     api = wandb.sdk.internal.internal_api.Api(
@@ -1181,7 +1181,7 @@ def test_launch_local_cuda_config(
     live_mock_server, test_settings, monkeypatch, mocked_fetchable_git_repo
 ):
     monkeypatch.setattr(
-        "wandb.sdk.launch.runner.local._run_entry_point",
+        "wandb.sdk.launch.runner.local_container._run_entry_point",
         lambda cmd, _: cmd,
     )
     api = wandb.sdk.internal.internal_api.Api(
@@ -1209,7 +1209,7 @@ def test_launch_cuda_prev_run_cuda(
     mock_cuda_run_info,
 ):
     monkeypatch.setattr(
-        "wandb.sdk.launch.runner.local._run_entry_point",
+        "wandb.sdk.launch.runner.local_container._run_entry_point",
         lambda cmd, _: cmd,
     )
     api = wandb.sdk.internal.internal_api.Api(
@@ -1236,7 +1236,7 @@ def test_launch_cuda_false_prev_run_cuda(
     mock_cuda_run_info,
 ):
     monkeypatch.setattr(
-        "wandb.sdk.launch.runner.local._run_entry_point",
+        "wandb.sdk.launch.runner.local_container._run_entry_point",
         lambda cmd, _: cmd,
     )
     api = wandb.sdk.internal.internal_api.Api(
@@ -1263,7 +1263,7 @@ def test_launch_cuda_config_false_prev_run_cuda(
     mock_cuda_run_info,
 ):
     monkeypatch.setattr(
-        "wandb.sdk.launch.runner.local._run_entry_point",
+        "wandb.sdk.launch.runner.local_container._run_entry_point",
         lambda cmd, _: cmd,
     )
     api = wandb.sdk.internal.internal_api.Api(
@@ -1346,7 +1346,7 @@ def test_launch_build_config_file(
     runner, mocked_fetchable_git_repo, test_settings, monkeypatch
 ):
     monkeypatch.setattr(
-        wandb.sdk.launch.runner.local.LocalContainerRunner,
+        wandb.sdk.launch.runner.local_container.LocalContainerRunner,
         "run",
         lambda *args, **kwargs: (args, kwargs),
     )
